requirements.txt
fastapi
jinja2
markdown
mermaid-py
pygments